Moved Renderer, Texture and TextureFactory from "core" to "base" directory
