+++ project/src/ts/base/world.ts
@@ -11,6 +11,7 @@ import WallTile from "../tiles/wall-tile";
 export default class World {
   public readonly SIZE: number = 20; // 20x20 world size
   public readonly CENTER: Vector = new Vector(this.SIZE / 2, this.SIZE / 2).floor();
+  public readonly ENTITY_CLICKABLE_RANGE: Vector = new Vector(3, 3);
 
   private _tiles: Tile[][];
   private _createdAt: number = Date.now();
@@ -54,6 +55,30 @@ export default class World {
   }
 
   public onWorldClicked (pos: Vector): boolean {
+    // TODO: FIX
+
+    // Const clickRange = new Vector(this.ENTITY_CLICKABLE_RANGE.x / 2, this.ENTITY_CLICKABLE_RANGE.y / 2);
+
+    // // Iterate through entities
+    // For (const entity of this._entities) {
+    //   // Entity was clicked
+    //   If ((entity.position.x - clickRange.x >= pos.x && entity.position.x + clickRange.x <= pos.x) &&
+    //       (entity.position.y - clickRange.y >= pos.y && entity.position.y + clickRange.y <= pos.y)) {
+    //     // Clicked entity is robot and hasn't started exploding yet
+    //     If (entity instanceof RobotEntity && !entity.hasStartedExploding) {
+    //       // Ensure entity stops moving
+    //       Entity.target = null;
+
+    //       // Make entity explode
+    //       Entity.explode();
+
+    //       Alert("EXPLODE");
+
+    //       Return true;
+    //     }
+    //   }
+    // }
+
     return false;
   }
