+++ project/src/ts/game.ts
@@ -1,7 +1,7 @@
 import GameLoop from "./core/game-loop";
 import Browser from "./browser/browser";
 import World from "./base/world";
-import Renderer from "./core/renderer";
+import Renderer from "./base/renderer";
 import Vector from "./core/vector";
 import TitleScreen from "./title-screen/title-screen";
 import Sound from "./base/sound";
